requirements.txt
langchain
requests
together
langchain-community
streamlit
python-dotenv
